# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "RAG Chatbot API for Physical AI & Humanoid Robotics Textbook",
    "description": "API for the RAG chatbot that answers questions about Physical AI and Humanoid Robotics curriculum",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/ingest": {
      "post": {
        "tags": [
          "ingest",
          "ingestion"
        ],
        "summary": "Ingest Content",
        "description": "Ingest a single content chunk into the vector database",
        "operationId": "ingest_content_api_v1_ingest_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IngestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ingest-batch": {
      "post": {
        "tags": [
          "ingest",
          "ingestion"
        ],
        "summary": "Ingest Batch Content",
        "description": "Ingest multiple content chunks in a batch",
        "operationId": "ingest_batch_content_api_v1_ingest_batch_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IngestBatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestBatchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ingest-directory": {
      "post": {
        "tags": [
          "ingest",
          "ingestion"
        ],
        "summary": "Ingest Directory Content",
        "description": "Ingest all content from a directory",
        "operationId": "ingest_directory_content_api_v1_ingest_directory_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IngestDirectoryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "IngestBatchRequest": {
        "properties": {
          "chunks": {
            "items": {
              "$ref": "#/components/schemas/IngestRequest"
            },
            "type": "array",
            "title": "Chunks"
          }
        },
        "type": "object",
        "required": [
          "chunks"
        ],
        "title": "IngestBatchRequest"
      },
      "IngestBatchResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "chunks_processed": {
            "type": "integer",
            "title": "Chunks Processed"
          },
          "success_count": {
            "type": "integer",
            "title": "Success Count"
          },
          "error_count": {
            "type": "integer",
            "title": "Error Count"
          },
          "results": {
            "items": {
              "type": "object"
            },
            "type": "array",
            "title": "Results"
          }
        },
        "type": "object",
        "required": [
          "status",
          "chunks_processed",
          "success_count",
          "error_count",
          "results"
        ],
        "title": "IngestBatchResponse"
      },
      "IngestDirectoryRequest": {
        "properties": {
          "directory_path": {
            "type": "string",
            "title": "Directory Path"
          },
          "recursive": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recursive",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "directory_path"
        ],
        "title": "IngestDirectoryRequest"
      },
      "IngestRequest": {
        "properties": {
          "module": {
            "type": "string",
            "title": "Module"
          },
          "week": {
            "type": "string",
            "title": "Week"
          },
          "chapter_title": {
            "type": "string",
            "title": "Chapter Title"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "file_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "File Path"
          }
        },
        "type": "object",
        "required": [
          "module",
          "week",
          "chapter_title",
          "content",
          "url"
        ],
        "title": "IngestRequest"
      },
      "IngestResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "chunk_id": {
            "type": "string",
            "title": "Chunk Id"
          },
          "token_count": {
            "type": "integer",
            "title": "Token Count"
          },
          "module": {
            "type": "string",
            "title": "Module"
          },
          "week": {
            "type": "string",
            "title": "Week"
          },
          "chapter_title": {
            "type": "string",
            "title": "Chapter Title"
          }
        },
        "type": "object",
        "required": [
          "status",
          "chunk_id",
          "token_count",
          "module",
          "week",
          "chapter_title"
        ],
        "title": "IngestResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}